django_project/requirements.txt
asgiref==3.7.2
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
gevent==23.9.0.post1
greenlet==2.0.2
pika==1.3.2
Pillow==10.0.0
psycopg2-binary==2.9.7
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
uuid==1.30
zope.event==5.0
zope.interface==6.0
